+++ src/actions/user.js
@@ -11,7 +11,9 @@ const MySwal = withReactContent(Swal)
 export const actFetchUsersRequest = () => {
     return async (dispatch) => {
         return await callApi('customer', 'GET', null, `Bearer ${getTokenEmployee()}`).then(res => {
-            dispatch(actFetchUser(res.data));
+            if(res!=null){
+                dispatch(actFetchUser(res.data));
+            }
         });
     }
 }
@@ -26,22 +28,24 @@ export const actFetchUser = (User) => {
 export const actAddUserRequest = (customer, history) => {
     return async () => {
         return await callApi('customer', 'POST', customer, `Bearer ${getTokenEmployee()}`).then(res => {
-            if (res.data.result === 1) {
-                MySwal.fire({
-                    icon: 'success',
-                    title: res.data.message,
-                    showConfirmButton: false,
-                    timer: 1500
-                })
-                // dispatch(actAddLineUser(res.data));
-                history.goBack()
-            }
-            else {
-                MySwal.fire({
-                    icon: 'error',
-                    title: 'Oops...',
-                    text: res.data.message
-                })
+            if(res!=null){
+                if (res.data.result === 1) {
+                    MySwal.fire({
+                        icon: 'success',
+                        title: res.data.message,
+                        showConfirmButton: false,
+                        timer: 1500
+                    })
+                    // dispatch(actAddLineUser(res.data));
+                    history.goBack()
+                }
+                else {
+                    MySwal.fire({
+                        icon: 'error',
+                        title: 'Oops...',
+                        text: res.data.message
+                    })
+                }
             }
         });
     }
@@ -50,21 +54,23 @@ export const actAddUserRequest = (customer, history) => {
 export const actUpdateUserRequest = (customer, history) => {
     return async (dispatch) => {
         return await callApi(`customer`, 'PUT', customer, `Bearer ${getTokenEmployee()}`).then(res => {
-            if (res.data.result === 1) {
-                MySwal.fire({
-                    icon: 'success',
-                    title: res.data.message,
-                    showConfirmButton: false,
-                    timer: 1500
-                })
-                history.goBack()
-                // dispatch(actUpdateLineUser(lineUser));
-            } else {
-                MySwal.fire({
-                    icon: 'error',
-                    title: 'Oops...',
-                    text: res.data.message
-                })
+            if(res!=null){
+                if (res.data.result === 1) {
+                    MySwal.fire({
+                        icon: 'success',
+                        title: res.data.message,
+                        showConfirmButton: false,
+                        timer: 1500
+                    })
+                    history.goBack()
+                    // dispatch(actUpdateLineUser(lineUser));
+                } else {
+                    MySwal.fire({
+                        icon: 'error',
+                        title: 'Oops...',
+                        text: res.data.message
+                    })
+                }
             }
         });
     }
@@ -73,20 +79,22 @@ export const actUpdateUserRequest = (customer, history) => {
 export const actDeleteUserRequest = (customerId) => {
     return async (dispatch) => {
         return await callApi(`customer/${customerId}`, 'DELETE', null, `Bearer ${getTokenEmployee()}`).then(res => {
-            if (res.data.result === 1) {
-                MySwal.fire({
-                    icon: 'success',
-                    title: res.data.message,
-                    showConfirmButton: false,
-                    timer: 1500
-                })
-                dispatch(actDeleteUser(customerId));
-            } else {
-                MySwal.fire({
-                    icon: 'error',
-                    title: 'Oops...',
-                    text: res.data.message
-                })
+            if(res!=null){
+                if (res.data.result === 1) {
+                    MySwal.fire({
+                        icon: 'success',
+                        title: res.data.message,
+                        showConfirmButton: false,
+                        timer: 1500
+                    })
+                    dispatch(actDeleteUser(customerId));
+                } else {
+                    MySwal.fire({
+                        icon: 'error',
+                        title: 'Oops...',
+                        text: res.data.message
+                    })
+                }
             }
         });
     }
